Add support for triggering an Assignment Rule
It doesn't appear to be documented anywhere, but I learned from Zoho
support that if you pass a `larid` param with the ID of a Zoho
Assignment Rule as a parameter, that assignment rule will be kicked off
in Zoho when you insert records via the API. I've only tested this out
with inserting Leads. Mileage may vary in other types of cases.